+++ Assets/OxGFrame/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [1.1.0] - 2022-07-25
+- Added USFrame (Unity Scene Manager) to controll unity scene, also supported AssetBundle.
+- Modified BundleMode (Bundle Stream Mode) only for build, can switch true or false from BundleConfig by define.
+
 ## [1.0.2] - 2022-07-12
 - Modified right-click feature to group CoreFrame and GSIFrame in OxGFrame path.